chatbot/requirements.txt
streamlit>=1.25.0
google-generativeai>=0.8.2
python-dotenv>=1.0.0